+++ src/CrudApp.Tests/Infrastructure/OpenApi/OpenApiTests.cs
@@ -11,7 +11,7 @@ public class OpenApiTests
         var fixture = await WebAppFixture.CreateAsync();
         var client = fixture.CreateHttpClient(fixture.InitialUserId);
         var response = await client.GetAsync("/swagger/v1/swagger.json");
-        Assert.Equal((int)HttpStatus.Ok, (int)response.StatusCode);
+        Assert.Equal(HttpStatus.Ok, (int)response.StatusCode);
         var openApiDoc = await response.Content.ReadAsStringAsync();
         await Verify(openApiDoc);
     }
